refactor: 패키지명 변경
RacingGame 이 Controller 의 역할을 하고 있다고 판단되었고, 이를 명확히 하기 위해서 package 명을 변경하였습니다.
